fix(stochastic): three bug fixes from architecture review
B1 — clear() vs sanitizeAfterRead() default divergence.
Extracted setDefaults() shared between both paths. Previously clear()
defaulted rhythmMode/melodyMode to Live and set _feel=50, while the
invalid-size recovery branch of sanitizeAfterRead() defaulted them to
Loop and skipped Feel entirely. A project with a corrupted size field
silently loaded as Loop instead of Live. Both paths now use the same
defaults; the difference is just whether _events is wiped (yes for clear,
yes for invalid-size, no for the normal sanitize path).

B2 — cycleTicks uint16 truncation.
Widened Cache::cycleTicks from uint16_t to uint32_t. The sum of per-cell
durations can reach 64 × 4095 = 262,080 ticks for dense long-duration
patterns, well past UINT16_MAX. The old saturating clamp prevented wrap
but fed the saturated value into Feel scaling, producing wrong Feel
adjustments at the high end. Now the true sum reaches Feel.

B3 — Mask rank in Repeat mode (documentation only).
Repeat replays the rank captured with _lastEvent at the slot the event
was last fired from. If permuteRhythm or mutateRhythm has reshuffled
ranks since capture, the replayed rank no longer matches the current
slot's stored rank. This is intentional under the Repeat contract
("frozen material" — replay what was captured, not what the slot looks
like now). Added an inline comment marking the choice so a future reader
doesn't try to "fix" it.

Sim + STM32 build clean. All 7 stochastic + queue suites green.

+++ src/apps/sequencer/engine/StochasticCache.h
@@ -108,7 +108,10 @@ struct Cache {
     uint8_t    parentCacheIdx[kMaxEventSlots];
 
     uint8_t    count;
-    uint16_t   cycleTicks;
+    // Sum of per-cell durations. Widened to 32 bits so dense long-duration
+    // patterns (worst case 64 × kMaxCellDuration = 262 080) reach Feel
+    // scaling without truncation.
+    uint32_t   cycleTicks;
 };
 
 // Build the cache from the sequence's events array.
